requirements/local.txt
# Local development dependencies go here
django-debug-toolbar
Sphinx
